Create README.md
Exploring the idea sending of small clips of data of variable size sent in explicit sequences as a way to obfuscate the transmission of a complex code being  from a client to  a server in a challenge-response type authentication system. One way transmission is key to this idea, keeping things dark, multiple iterations means getting it right once is not enough, you need to get it right every time. I am looking forward to constructive critiques from you all.