Create auth.py
This is for repository is for  research purposes. Secret key would be hidden, and print statements would not be included in a real world system.
Run the auth.py file first to initiate the flask server, then run the req.py in a separate terminal. Both scripts need to run at the same time